requirements.txt
fastapi[standard]>=0.110.0
pydantic>=2.6.4
uvicorn>=0.29.0
google-genai>=0.3.0
python-dotenv>=1.0.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
